backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pypdf==4.0.1
spacy==3.7.4
google-generativeai==0.3.2
nltk==3.8.1
langchain==0.1.9
python-multipart==0.0.9
pydantic==2.6.1
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
httpx==0.26.0
pytest==8.0.1
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 